+++ src/main/java/study/racingcar/GamePlay.java
@@ -1,5 +1,7 @@
 package study.racingcar;
 
+import study.racingcar.controller.GameController;
+
 public class GamePlay {
 	public static void main(String[] args) {
 		GameController.start();
