+++ src/main/java/study/racingcar/domain/Cars.java
@@ -1,4 +1,4 @@
-package study.racingcar;
+package study.racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
