+++ src/main/java/study/racingcar/domain/Game.java
@@ -1,4 +1,4 @@
-package study.racingcar;
+package study.racingcar.domain;
 
 public class Game {
