+++ src/test/java/study/racingcar/CarStatusTest.java
@@ -4,6 +4,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import study.racingcar.domain.Cars;
+import study.racingcar.domain.Game;
+import study.racingcar.domain.GameStatus;
+
 public class CarStatusTest {
 
 	private Cars cars;
