+++ src/test/java/study/racingcar/ResultViewTest.java
@@ -3,6 +3,8 @@ package study.racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import study.racingcar.view.ResultView;
+
 public class ResultViewTest {
 
 	@Test
